Move Calc into dimension.rs
